Move generated files into ./generated folder
