+++ packages/iov-bns/package.json
@@ -25,12 +25,12 @@
     "test": "yarn build-or-skip && yarn test-node",
     "prebuild": "yarn format && yarn lint",
     "move-types": "shx rm -r ./types/* && shx mv build/types/* ./types && shx rm ./types/*.spec.d.ts",
-    "build": "shx rm -rf ./build && tsc && shx cp ./src/codecimpl.js ./build && shx cp ./src/codecimpl.d.ts ./build/types && yarn move-types",
+    "build": "shx rm -rf ./build && tsc && shx mkdir -p build/generated && shx cp ./src/generated/*.js ./build/generated && shx mkdir -p ./build/types/generated && shx cp ./src/generated/*.d.ts ./build/types/generated && yarn move-types",
     "build-or-skip": "[ -n \"$SKIP_BUILD\" ] || yarn build",
     "pack-web": "yarn build-or-skip && webpack --mode development --config webpack.web.config.js",
     "find-proto": "find ./weave -name '*.proto' -not -path '*/vendor/*' -not -path '*/examples/*' -not -path '*/cmd/bcpd/*'",
-    "pack-proto": "pbjs -t static-module -w commonjs -o src/codecimpl.js `yarn -s find-proto`",
-    "define-proto": "pbts src/codecimpl.js -o src/codecimpl.d.ts",
+    "pack-proto": "pbjs -t static-module -w commonjs -o src/generated/codecimpl.js `yarn -s find-proto`",
+    "define-proto": "pbts src/generated/codecimpl.js -o src/generated/codecimpl.d.ts",
     "protoc": "./scripts/weave.sh && yarn pack-proto && yarn define-proto && yarn format"
   },
   "dependencies": {
